Add comprehensive test module
