+++ golem-rib/src/interpreter/rib_interpreter.rs
@@ -1874,347 +1874,6 @@ mod interpreter_tests {
         }
     }
 
-    // The test consist of all possibilities, such as no-arg functions, unit function, no-arg unit functions,
-    // simple functions, pattern matches, if-else condition etc
-    mod complex_rib_test {
-        use crate::interpreter::rib_interpreter::interpreter_tests::internal;
-        use crate::{compiler, Expr};
-        use golem_wasm_ast::analysis::{
-            AnalysedType, TypeList, TypeOption, TypeResult, TypeStr, TypeU64, TypeU8,
-        };
-        use std::collections::HashMap;
-
-        #[tokio::test]
-        async fn test_interpreter_complex_rib() {
-            let expr = r#"
-              let input = { body : { id: "bId", name: "bName", titles: request.body.titles, address: request.body.address } };
-              let optional_result = function-option(input);
-              let id: str = request.body.id;
-
-              let a = match optional_result {
-                some(value) => "personal-id",
-                none => id
-               };
-
-              let b = match optional_result {
-                some(value) => "personal-id",
-                none =>  input.body.titles[1]
-               };
-
-              let c = match optional_result {
-                some(value) => "personal-id",
-                none =>  request.body.address.street
-              };
-
-              let authorisation: str = request.headers.authorisation;
-              let success: u64 = 200;
-              let failure: u64 = 401;
-
-              let d = if authorisation == "admin" then success else failure;
-
-              function-no-arg-unit();
-
-              let unit_result = function-unit(input);
-
-              let ok_result = function-ok(input);
-
-              let e = if id == "foo" then "bar" else match ok_result {
-                ok(value) => value,
-                err(msg) => "empty"
-               };
-
-              let err_result = function-err(input);
-
-              let f = if id == "foo" then "bar" else match err_result { ok(value) => value, err(msg) => msg };
-
-              let ok_record_result = function-ok-record(input);
-
-              let g = match ok_record_result {
-                ok(success_rec) => success_rec.b[0],
-                err(failure_rec) => "${failure_rec.d}"
-              };
-
-              let err_record_result = function-err-record(input);
-
-              let h = match err_record_result { ok(success_rec) => success_rec.b[0], err(failure_rec) => "${failure_rec.d}" };
-              let i = match err_record_result { ok(_) => 1u8, err(_) => 0u8 };
-              let j = match ok_record_result { ok(_) => 1u8, err(_) => 0u8 };
-
-              let variant_result = function-variant(input);
-
-              let k = match variant_result {
-                 process-user(value) => value,
-                 register-user(value) => "${value}",
-                 validate => "validated"
-              };
-
-              let processed = process-user("jon");
-              let registered = register-user(1);
-
-              let m = match processed {
-                process-user(name) => name,
-                register-user(value) => "${value}",
-                validate => "validated"
-              };
-
-              let n = match registered {
-                process-user(name) => name,
-                register-user(value) => "${value}",
-                validate => "validated"
-              };
-
-              { a : a, b : b, c: c, d: d, e: e, f: f, g: g, h: h, i: i, j: j, k: k, m: m, n: n}
-        "#;
-
-            let expr = Expr::from_text(expr).unwrap();
-
-            let optional_result_type = AnalysedType::Option(TypeOption {
-                inner: Box::new(AnalysedType::Str(TypeStr)),
-            });
-
-            let ok_err_result_type = AnalysedType::Result(TypeResult {
-                ok: Some(Box::new(AnalysedType::Str(TypeStr))),
-                err: Some(Box::new(AnalysedType::Str(TypeStr))),
-            });
-
-            let ok_err_record_result_type = AnalysedType::Result(TypeResult {
-                ok: Some(Box::new(internal::analysed_type_record(vec![
-                    ("a", AnalysedType::Str(TypeStr)),
-                    (
-                        "b",
-                        AnalysedType::List(TypeList {
-                            inner: Box::new(AnalysedType::Str(TypeStr)),
-                        }),
-                    ),
-                ]))),
-                err: Some(Box::new(internal::analysed_type_record(vec![
-                    ("c", AnalysedType::Str(TypeStr)),
-                    ("d", AnalysedType::Str(TypeStr)),
-                ]))),
-            });
-
-            let variant_result_type = internal::get_analysed_type_variant();
-
-            let optional_result = internal::get_type_annotated_value(&optional_result_type, "none");
-            let ok_result =
-                internal::get_type_annotated_value(&ok_err_result_type, r#"ok("success")"#);
-            let err_result =
-                internal::get_type_annotated_value(&ok_err_result_type, r#"err("failure")"#);
-            let ok_record_result = internal::get_type_annotated_value(
-                &ok_err_record_result_type,
-                r#"ok({a : "foo", b: ["bar", "baz"]})"#,
-            );
-
-            let err_record_result = internal::get_type_annotated_value(
-                &ok_err_record_result_type,
-                r#"err({c : "bar", d: "fuuz"})"#,
-            );
-
-            let variant_result =
-                internal::get_type_annotated_value(&variant_result_type, r#"validate"#);
-
-            let record_input_type = internal::analysed_type_record(vec![
-                (
-                    "headers",
-                    internal::analysed_type_record(vec![(
-                        "authorisation",
-                        AnalysedType::Str(TypeStr),
-                    )]),
-                ),
-                (
-                    "body",
-                    internal::analysed_type_record(vec![
-                        ("id", AnalysedType::Str(TypeStr)),
-                        ("name", AnalysedType::Str(TypeStr)),
-                        (
-                            "titles",
-                            AnalysedType::List(TypeList {
-                                inner: Box::new(AnalysedType::Str(TypeStr)),
-                            }),
-                        ),
-                        (
-                            "address",
-                            internal::analysed_type_record(vec![
-                                ("street", AnalysedType::Str(TypeStr)),
-                                ("city", AnalysedType::Str(TypeStr)),
-                            ]),
-                        ),
-                    ]),
-                ),
-            ]);
-
-            let input_value = internal::get_type_annotated_value(
-                &record_input_type,
-                r#" {
-                   headers : {
-                      authorisation : "admin"
-                   },
-                   body : {
-
-
-                           id: "bId",
-                           name: "bName",
-                           titles: [
-                             "bTitle1", "bTitle2"
-                           ],
-                           address: {
-                             street: "bStreet",
-                             city: "bCity"
-                           }
-
-                    }
-                 }"#,
-            );
-
-            let mut function_option_metadata = internal::get_component_metadata(
-                "function-option",
-                vec![record_input_type.clone()],
-                Some(optional_result_type.clone()),
-            );
-
-            let function_no_arg_unit_metadata =
-                internal::get_component_metadata("function-no-arg-unit", vec![], None);
-
-            let function_unit_metadata =
-                internal::get_component_metadata("function-unit", vec![record_input_type.clone()], None);
-
-            let function_ok_metadata = internal::get_component_metadata(
-                "function-ok",
-                vec![record_input_type.clone()],
-                Some(ok_err_result_type.clone()),
-            );
-
-            let function_err_metadata = internal::get_component_metadata(
-                "function-err",
-                vec![record_input_type.clone()],
-                Some(ok_err_result_type),
-            );
-
-            let function_ok_record_metadata = internal::get_component_metadata(
-                "function-ok-record",
-                vec![record_input_type.clone()],
-                Some(ok_err_record_result_type.clone()),
-            );
-
-            let function_err_record_metadata = internal::get_component_metadata(
-                "function-err-record",
-                vec![record_input_type.clone()],
-                Some(ok_err_record_result_type),
-            );
-
-            let function_variant_metadata = internal::get_component_metadata(
-                "function-variant",
-                vec![record_input_type.clone()],
-                Some(variant_result_type),
-            );
-
-            function_option_metadata.extend(function_no_arg_unit_metadata);
-            function_option_metadata.extend(function_unit_metadata);
-            function_option_metadata.extend(function_ok_metadata);
-            function_option_metadata.extend(function_err_metadata);
-            function_option_metadata.extend(function_ok_record_metadata);
-            function_option_metadata.extend(function_err_record_metadata);
-            function_option_metadata.extend(function_variant_metadata);
-
-            let compiled = compiler::compile(&expr, &function_option_metadata).unwrap();
-
-            let interpreter_input = HashMap::from_iter(vec![("request".to_string(), input_value)]);
-
-            let functions_and_result = HashMap::from_iter(vec![
-                (
-                    internal::FunctionName("function-option".to_string()),
-                    Some(optional_result),
-                ),
-                (
-                    internal::FunctionName("function-ok".to_string()),
-                    Some(ok_result),
-                ),
-                (
-                    internal::FunctionName("function-err".to_string()),
-                    Some(err_result),
-                ),
-                (
-                    internal::FunctionName("function-ok-record".to_string()),
-                    Some(ok_record_result),
-                ),
-                (
-                    internal::FunctionName("function-err-record".to_string()),
-                    Some(err_record_result),
-                ),
-                (
-                    internal::FunctionName("function-variant".to_string()),
-                    Some(variant_result),
-                ),
-                (
-                    internal::FunctionName("function-no-arg-unit".to_string()),
-                    None,
-                ),
-                (internal::FunctionName("function-unit".to_string()), None),
-            ]);
-            let mut rib_executor =
-                internal::dynamic_test_interpreter(functions_and_result, interpreter_input);
-            let result = rib_executor.run(compiled.byte_code).await.unwrap();
-
-            let expected_result = internal::get_type_annotated_value(
-                &internal::analysed_type_record(vec![
-                    ("a", AnalysedType::Str(TypeStr)),
-                    ("b", AnalysedType::Str(TypeStr)),
-                    ("c", AnalysedType::Str(TypeStr)),
-                    ("d", AnalysedType::U64(TypeU64)),
-                    ("e", AnalysedType::Str(TypeStr)),
-                    ("f", AnalysedType::Str(TypeStr)),
-                    ("g", AnalysedType::Str(TypeStr)),
-                    ("h", AnalysedType::Str(TypeStr)),
-                    ("i", AnalysedType::U8(TypeU8)),
-                    ("j", AnalysedType::U8(TypeU8)),
-                    ("k", AnalysedType::Str(TypeStr)),
-                    ("m", AnalysedType::Str(TypeStr)),
-                    ("n", AnalysedType::Str(TypeStr)),
-                ]),
-                r#" { a : "bId", b : "bTitle2", c : "bStreet", d: 200, e: "success", f: "failure", g: "bar", h : "fuuz", i: 0, j: 1, k: "validated", m:"jon", n: "1" }"#,
-            );
-            assert_eq!(result.get_val().unwrap(), expected_result);
-        }
-
-
-        mod data {
-            use golem_wasm_ast::analysis::{AnalysedType, TypeList, TypeStr};
-            use crate::interpreter::rib_interpreter::interpreter_tests::internal;
-
-            pub(crate) fn record_input_type() -> AnalysedType {
-                internal::analysed_type_record(vec![
-                    (
-                        "headers",
-                        internal::analysed_type_record(vec![(
-                            "authorisation",
-                            AnalysedType::Str(TypeStr),
-                        )]),
-                    ),
-                    (
-                        "body",
-                        internal::analysed_type_record(vec![
-                            ("id", AnalysedType::Str(TypeStr)),
-                            ("name", AnalysedType::Str(TypeStr)),
-                            (
-                                "titles",
-                                AnalysedType::List(TypeList {
-                                    inner: Box::new(AnalysedType::Str(TypeStr)),
-                                }),
-                            ),
-                            (
-                                "address",
-                                internal::analysed_type_record(vec![
-                                    ("street", AnalysedType::Str(TypeStr)),
-                                    ("city", AnalysedType::Str(TypeStr)),
-                                ]),
-                            ),
-                        ]),
-                    ),
-                ])
-            }
-        }
-    }
-
     mod internal {
         use crate::interpreter::env::InterpreterEnv;
         use crate::interpreter::stack::InterpreterStack;
@@ -2562,57 +2221,6 @@ mod interpreter_tests {
             wasm_wave_str: &str,
         ) -> TypeAnnotatedValue {
             golem_wasm_rpc::type_annotated_value_from_str(analysed_type, wasm_wave_str).unwrap()
-        }
-
-        #[derive(Clone, Hash, PartialEq, Eq)]
-        pub(crate) struct FunctionName(pub(crate) String);
-
-        pub(crate) fn dynamic_test_interpreter(
-            functions_and_result: HashMap<FunctionName, Option<TypeAnnotatedValue>>,
-            interpreter_env_input: HashMap<String, TypeAnnotatedValue>,
-        ) -> Interpreter {
-            Interpreter {
-                stack: InterpreterStack::default(),
-                env: InterpreterEnv::from(
-                    interpreter_env_input,
-                    dynamic_worker_invoke(functions_and_result),
-                ),
-            }
-        }
-
-        fn dynamic_worker_invoke(
-            functions_and_result: HashMap<FunctionName, Option<TypeAnnotatedValue>>,
-        ) -> RibFunctionInvoke {
-            let value = functions_and_result.clone();
-
-            Arc::new(move |a, _| {
-                Box::pin({
-                    let value = value.get(&FunctionName(a)).cloned().flatten();
-                    let analysed_type = value.clone().map(|x| AnalysedType::try_from(&x).unwrap());
-
-                    async move {
-                        let analysed_type = analysed_type.clone();
-                        let value = value.clone();
-
-                        if let Some(value) = value {
-                            Ok(TypeAnnotatedValue::Tuple(TypedTuple {
-                                typ: vec![golem_wasm_ast::analysis::protobuf::Type::from(
-                                    &analysed_type.unwrap(),
-                                )],
-                                value: vec![golem_wasm_rpc::protobuf::TypeAnnotatedValue {
-                                    type_annotated_value: Some(value),
-                                }],
-                            }))
-                        } else {
-                            // Representing Unit
-                            Ok(TypeAnnotatedValue::Tuple(TypedTuple {
-                                typ: vec![],
-                                value: vec![],
-                            }))
-                        }
-                    }
-                })
-            })
         }
 
         pub(crate) fn static_test_interpreter(
